internal: type adjusts
