+++ generators/client/generator.ts
@@ -29,11 +29,12 @@ import { addEnumerationFiles } from './entity-files.js';
 import { writeFiles as writeCommonFiles } from './files-common.js';
 import { askForClientTheme, askForClientThemeVariant } from './prompts.js';
 import { filterEntitiesAndPropertiesForClient } from './support/filter-entities.js';
+import type { Application as ClientApplication, Entity as ClientEntity } from './types.d.ts';
 
 const { ANGULAR, NO: CLIENT_FRAMEWORK_NO } = clientFrameworkTypes;
 const { CYPRESS } = testFrameworkTypes;
 
-export default class JHipsterClientGenerator extends BaseApplicationGenerator {
+export default class ClientGenerator extends BaseApplicationGenerator<ClientEntity, ClientApplication<ClientEntity>> {
   async beforeQueue() {
     if (!this.fromBlueprint) {
       await this.composeWithBlueprints();
