+++ generators/client/index.ts
@@ -19,4 +19,4 @@
 export { default } from './generator.js';
 export { default as command } from './command.js';
 export { files as commonFiles } from './files-common.js';
-export type { FrontendApplication as Application, Entity, Field, Relationship } from './types.js';
+export type { Application, Entity, Field, Relationship } from './types.js';
