+++ generators/javascript/generators/eslint/support/tasks.ts
@@ -18,18 +18,18 @@
  */
 import { asWritingTask } from '../../../../base-application/support/index.js';
 import { clientRootTemplatesBlock } from '../../../../client/support/files.js';
-import type { Application as JavascriptApplication, Entity as JavascriptEntity } from '../../../entity.d.ts';
+import type { Application as JavascriptApplication, Entity as JavascriptEntity } from '../../../types.d.ts';
 
-export const writeEslintClientRootConfigFile = asWritingTask<JavascriptEntity, JavascriptApplication>(async function writingEslintFile({
-  application,
-}) {
-  await this.writeFiles({
-    blocks: [
-      clientRootTemplatesBlock({
-        templates: [{ sourceFile: 'eslint.config.js.jhi', destinationFile: ctx => `${ctx.eslintConfigFile}.jhi` }],
-      }),
-    ],
-    context: application,
-    rootTemplatesPath: this.fetchFromInstalledJHipster('javascript/generators/eslint/templates'),
-  });
-});
+export const writeEslintClientRootConfigFile = asWritingTask<JavascriptEntity, JavascriptApplication<JavascriptEntity>>(
+  async function writingEslintFile({ application }) {
+    await this.writeFiles({
+      blocks: [
+        clientRootTemplatesBlock({
+          templates: [{ sourceFile: 'eslint.config.js.jhi', destinationFile: ctx => `${ctx.eslintConfigFile}.jhi` }],
+        }),
+      ],
+      context: application,
+      rootTemplatesPath: this.fetchFromInstalledJHipster('javascript/generators/eslint/templates'),
+    });
+  },
+);
